requirements.txt
requests>=2.31.0
pydantic>=2.0.0
pyyaml>=6.0.0
click>=8.1.0
rich>=13.0.0
websocket-client>=1.6.0
aiohttp>=3.9.0
schedule>=1.2.0
discord-webhook>=1.2.0
tenacity>=8.2.0
python-dotenv>=1.0.0

# Google Sheets dashboard
gspread>=5.12.0
google-auth>=2.25.0

# Test dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
responses>=0.24.0
